+++ CHANGELOG.rst
@@ -8,6 +8,10 @@ Fixed:
 ^^^^^^
 - Fixed websocket handler to match the data_type returned from BE for local data analysis
 
+Changed:
+^^^^^^^
+- Changed Heatmap default to a range of 0-1 Hz and changed default metric to Twitch Frequency
+
 
 1.3.1 (2022-12-13)
 ------------------
@@ -35,7 +39,6 @@ Changed:
 - Error message now reads: "Please turn the instrument off, unplug from the PC, and then wait 10 seconds before attempting to use again"
 - Allow pasting barcodes into barcode text box when manual entry is enabled
 - Renamed ``recording_snapshot`` websocket handler to ``recording_snapshot_data``
-- Changed Heatmap default to a range of 0-1 Hz and changed default metric to Twitch Frequency
 
 Fixed:
 ^^^^^^